+++ cmd/monitor/collector/main.go
@@ -25,7 +25,6 @@ import (
 
 	// providers
 	_ "github.com/erda-project/erda/modules/core/monitor/collector"
-	_ "github.com/erda-project/erda/modules/core/monitor/collector/v2"
 )
 
 //go:generate sh -c "cd ${PROJ_PATH} && go generate -v -x github.com/erda-project/erda/modules/monitor/core/collector"
