+++ cmd/monitor/streaming/main.go
@@ -20,7 +20,6 @@ import (
 
 	// modules
 	_ "github.com/erda-project/erda/modules/core/monitor/log/storage"
-	_ "github.com/erda-project/erda/modules/core/monitor/log/storage/v2"
 	_ "github.com/erda-project/erda/modules/core/monitor/metric/storage"
 	_ "github.com/erda-project/erda/modules/monitor/alert/storage/alert-record"
 	_ "github.com/erda-project/erda/modules/monitor/apm/browser"
