+++ modules/core/monitor/log/storage/consumer_test.go
@@ -11,67 +11,61 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package storagev2
+package storage
 
 import (
 	"testing"
 	"time"
 
-	"github.com/alecthomas/assert"
-
-	"github.com/erda-project/erda-proto-go/core/monitor/log/storage/pb"
 	logmodule "github.com/erda-project/erda/modules/core/monitor/log"
-	"github.com/golang/protobuf/proto"
+	"github.com/stretchr/testify/assert"
 )
 
-func Test_provider_processLogV2(t *testing.T) {
+func Test_provider_processLog(t *testing.T) {
 	mp := mockProvider()
 	ass := assert.New(t)
 
 	// default
-	log := &pb.Log{
-		Id:        "aaa",
+	log := &logmodule.Log{
+		ID:        "aaa",
 		Source:    "container",
 		Stream:    "",
 		Offset:    0,
 		Timestamp: 0,
 		Content:   "",
 		Tags:      nil,
-		Labels:    nil,
 	}
-	mp.processLogV2(log)
+	mp.processLog(log)
 	ass.Equal("INFO", log.Tags["level"])
 	ass.Equal("stdout", log.Stream)
 
 	// upper level
-	log = &pb.Log{
-		Id:        "aaa",
+	log = &logmodule.Log{
+		ID:        "aaa",
 		Source:    "container",
 		Stream:    "",
 		Offset:    0,
 		Timestamp: 0,
 		Content:   "",
 		Tags:      map[string]string{"level": "error"},
-		Labels:    nil,
 	}
-	mp.processLogV2(log)
+	mp.processLog(log)
 	ass.Equal("ERROR", log.Tags["level"])
 	ass.Equal("stdout", log.Stream)
 
 	// id keys
 	mp.Cfg.Output.IDKeys = []string{"id_key1"}
-	log = &pb.Log{
-		Id:        "aaa",
+	log = &logmodule.Log{
+		ID:        "aaa",
 		Source:    "container",
 		Stream:    "",
 		Offset:    0,
 		Timestamp: 0,
 		Content:   "",
 		Tags:      map[string]string{"id_key1": "bbb"},
-		Labels:    nil,
 	}
-	mp.processLogV2(log)
-	ass.Equal("bbb", log.Id)
+	mp.processLog(log)
+	ass.Equal("bbb", log.ID)
 }
 
 func Test_provider_invokeV2(t *testing.T) {
@@ -80,33 +74,29 @@ func Test_provider_invokeV2(t *testing.T) {
 	mp.output = mw
 	ass := assert.New(t)
 
-	log := &pb.Log{
-		Id:        "aaa",
+	log := &logmodule.Log{
+		ID:        "aaa",
 		Source:    "container",
 		Stream:    "stdout",
 		Offset:    1024,
 		Timestamp: time.Now().UnixNano(),
 		Content:   "hello world",
 		Tags:      map[string]string{"level": "INFO"},
-		Labels:    nil,
 	}
 
-	lb := &pb.LogBatch{
-		Logs: []*pb.Log{log},
-	}
-	value, err := proto.Marshal(lb)
+	value, err := json.Marshal(log)
 	ass.Nil(err)
-	err = mp.invokeV2(nil, value, nil, time.Now())
+	err = mp.invoke(nil, value, nil, time.Now())
 	ass.Nil(err)
 	ass.Equal(&logmodule.LogMeta{
-		ID:     log.Id,
+		ID:     log.ID,
 		Source: log.Source,
 		Tags:   log.Tags,
 	}, mw.datas[0])
-	proto.Equal(log, mw.datas[1].(*pb.Log))
+	ass.Equal(log, mw.datas[1].(*logmodule.Log))
 
 	// bad value
-	err = mp.invokeV2(nil, []byte(`bad value`), nil, time.Now())
+	err = mp.invoke(nil, []byte(`bad value`), nil, time.Now())
 	ass.Error(err)
 }
